requirements.txt
psycopg2-binary
pandas
numpy
yfinance
requests
python-dotenv
ratelimit
fastapi
uvicorn[standard]
pytz
pandas-market-calendars 
pydantic 